Move routes to lib folder
