+++ frontend/components/pokemons/ListItem.tsx
@@ -1,6 +1,6 @@
 import { NarrowPokemonFieldsFragment } from '@/__generated__/graphql'
 import { PokemonDescription } from '@/components/common/PokemonDescription'
-import { getPokemonDetailRoute } from '@/components/common/routes'
+import { getPokemonDetailRoute } from '@/lib/routes'
 import styles from '@/styles/pokemons.module.scss'
 import Image from 'next/image'
 import Link from 'next/link'
